refactor: split webhook monitor into components
